baseline models to get benchmark results for advanced models

Get results for the following simple forecasting methods:

Naive forecast (naive)

Seasonal naive forecast (snaive)

Mean forecast (meanf)

Simple exponential smoothing forecasts (ses)

Exponential smoothing with Holt's trend (holt)

forecast_baseline(data, test_size = NULL, acc_measure)

Arguments

data
Univariate time series object

test_size
Numeric vector of length one specifying length of test set (along with forecasting horizon). NULL by default which yield 2 times the frequency of the time series

acc_measure
Accuracy indicator. Valid options are "ME", "RMSE", "MAE", "MPE", "MAPE", "MASE" and "ACF1"

Value

list of forecasting methods with data.frame for accuracy indicators

Examples

# Calculate MAE and RMSE of simple forecasting models for Apple FCF
apple <- tsRNN::ts_apple
forecast_baseline(apple, acc_measure = c("MAE", "RMSE"))

#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 

#> $fc_naive
#>                   MAE     RMSE
#> Training set 1526.402 3335.029
#> Test set     4100.125 6190.171
#> 
#> $fc_snaive
#>                   MAE     RMSE
#> Training set 1030.414 1962.797
#> Test set     2025.375 2266.507
#> 
#> $fc_mean
#>                   MAE      RMSE
#> Training set  5436.22  6650.121
#> Test set     12091.95 13032.294
#> 
#> $fc_ses
#>                   MAE     RMSE
#> Training set 1312.167 2802.376
#> Test set     3761.186 4904.246
#> 
#> $fc_holt
#>                   MAE     RMSE
#> Training set 1376.715 2738.735
#> Test set     3626.459 4896.691
#> 

# Change forecast horizom to 4 years
forecast_baseline(apple, test_size = 16, acc_measure = c("MAE", "RMSE"))

#> $fc_naive
#>                   MAE     RMSE
#> Training set 1110.851 2613.241
#> Test set     6110.375 7967.478
#> 
#> $fc_snaive
#>                    MAE     RMSE
#> Training set  989.2276 1974.340
#> Test set     1457.3750 1665.448
#> 
#> $fc_mean
#>                    MAE      RMSE
#> Training set  4524.244  5849.772
#> Test set     12757.006 13743.553
#> 
#> $fc_ses
#>                   MAE     RMSE
#> Training set 1008.348 2322.552
#> Test set     3850.357 5445.901
#> 
#> $fc_holt
#>                   MAE     RMSE
#> Training set 1055.977 2287.625
#> Test set     5499.393 5818.100
#>