Train Recurrent Neural Network with Gated Recurrent Unit (GRU) or Long-Short Term Memory (LSTM) using Keras framework

Train Recurrent Neural Network with Gated Recurrent Unit (GRU) or Long-Short Term Memory (LSTM) using Keras framework

keras_rnn(
  X,
  Y,
  model_type,
  tsteps,
  n_epochs = 200,
  n_units = 32,
  loss = "mse",
  metrics = NULL,
  dropout_in_test = FALSE,
  optimizer = optimizer_rmsprop(),
  dropout = 0,
  recurrent_dropout = 0,
  live_plot = FALSE
)

Arguments

X
list of "train", "val", and "test" with 3D (keras) arrays

Y
list of "train", "val", and "test" with 2D (keras) arrays

model_type
One of "simple", "gru" and "lstm"

tsteps
number of time steps for keras input shape

n_epochs
default 200

n_units
32 (currently fixed)

loss
default "mse"

metrics
default NULL

dropout_in_test
apply dropout during training only (default) or during testing also? Required for dropout-based prediction intervals (bayesian RNN)

optimizer
from keras, e.g. optimizer_rmsprop()

dropout
dropout rate

recurrent_dropout
Dropout rate applied to reccurent layer. Default 0

live_plot
plot loss and validation metric during training? False by default

Value

evaluation scores for training, validation and test set