Prediction and evaluation for baseline models including cross validation

Prediction and evaluation for baseline models including cross validation

predict_baselines(
  data,
  cv_setting,
  col_id = NULL,
  col_date = "index",
  col_value = "value",
  transform = NULL,
  frequency = 4,
  multiple_h = NULL
)

Arguments

data
data.table object

cv_setting
cross valudation settings. List Requiring

periods_train

periods_val

periods_test

skip_span

col_id
Optional ID column in data, default to "ticker"

col_date
Date column in data, default to "index"

col_value
Value column in data, default to "value"

transform
One of NULL, "normalize" and "box" for Box-Cox transformation

frequency
time series frequency, e.g. 4 for quarters and 12 for months

multiple_h
NULL if forecast horizon equals cv_setting$n_test, else named list of forecast horizons for accuracy measures

Value

list with accuracy and forecasts (point forecast and PI)