Tune recurrent neural network with Keras functional API and Bayes Optimization to select best performing model

Tune recurrent neural network with Keras functional API and Bayes Optimization to select best performing model

tune_keras_rnn_bayesoptim(
  data,
  model_type,
  cv_setting,
  tuning_bounds,
  col_id = NULL,
  col_date = "index",
  col_value = "value",
  save = NULL,
  save_id = NULL
)

Arguments

data
Univariate time series (data.frame) with date and value column, specified in col_date and col_value

model_type
One of "simple", "gru" or "lstm"

cv_setting
list of "periods_train", "periods_val", "periods_test" and "skip_span" for rolling_origin

tuning_bounds
list of tuning parameters - see section "Tuning Bounds"

col_id
Optional ID column in data, default to "ticker"

col_date
Date column in data, default to "index"

col_value
Value column in data, default to "value"

save
Automatically save tuning results? Specify NULL if not or character vector with path to directory for yes

save_id
optional id for model filename

Value

list of Bayes Optimization results per split

Tuning Bounds

The following parameters are (currently) available for tuning.

lag_1 (integer(2)) lower bounds

lag_2 (integer(2)) upper bounds

n_units (integer(2)) lower and upper bound for rnn units (cells)

n_epochs (integer(2)) lower and upper bound for epochs

optimizer_type (integer(2)) lower and upper bound for optimizer: 1 = "rmsprop", 2 = "adam", 3 = "adagrad"

dropout = (numeric(2)) lower and upper bound for dropout rate

recurrent_dropout = (numeric(2)) lower and upper bound for recurrent dropout rate

learning_rate = (numeric(2)) lower and upper bound for learning rate